Moving files as per PR comments
